% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_ResidualsModified.R
\name{CalcPolyAResidualsPolyA}
\alias{CalcPolyAResidualsPolyA}
\title{Calculate PolyA Residuals}
\usage{
CalcPolyAResidualsPolyA(
  object,
  assay = "polyA",
  features = NULL,
  background = NULL,
  gene.names = "Gene_Symbol",
  min.counts.background = 5,
  min.variance = 0.1,
  sample.n = 1000,
  do.scale = FALSE,
  do.center = FALSE,
  residuals.max = NULL,
  residuals.min = NULL,
  number.bins = 30,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object containing a polyAsiteAssay}

\item{assay}{Name of polyAsiteAssay to be used in calculating polyAresiduals}

\item{features}{Features to include in calculation of polyA residuals.
Default is to use all features.}

\item{background}{Identity of cells to use as background.
Default is to use all cells as a background.}

\item{gene.names}{Column containing the gene where each polyA site is annotated.
Default is symbol.}

\item{min.counts.background}{Features with at least this many counts in the background cells are included in calculation}

\item{min.variance}{Sets minimum variance. Default is 0.1.}

\item{sample.n}{Max number of observations to sample in each bin when performing regularization. Default is 1000.}

\item{do.scale}{Return the scaled residuals. Default is FALSE.}

\item{do.center}{Return the centered residuals. Default is FALSE.}

\item{residuals.max}{Clip residuals above this value. Default is NULL (no clipping).}

\item{residuals.min}{Clip residuals below this value. Default is NULL (no clipping).}

\item{number.bins}{Number of bins used to perform regularization. Default is 30.}

\item{verbose}{Print messages.}
}
\value{
Returns a Seurat object with polyAresiduals assay
}
\description{
Calculate PolyA Residuals
}
\examples{
\dontrun{
seurat_obj <- CalcPolyAResidualsPolyA(
  seurat_obj,
  assay = "polyA",
  gene.names = "Gene_Symbol",
  background = "Control",
  min.counts.background = 5
)
}

}
\concept{residuals}
