% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_PolyAAnalysis.R
\name{CleanPolyAMetadata}
\alias{CleanPolyAMetadata}
\title{Clean PolyA Metadata by Removing Duplicate and Positional Columns}
\usage{
CleanPolyAMetadata(
  obj,
  remove_coords = TRUE,
  coord_cols = c("start", "end", "width", "strand", "seqnames"),
  verbose = TRUE
)
}
\arguments{
\item{obj}{Seurat object with polyA assay}

\item{remove_coords}{Remove coordinate columns? Default TRUE}

\item{coord_cols}{Vector of coordinate column names to remove}

\item{verbose}{Print messages? Default TRUE}
}
\value{
Seurat object with cleaned polyA metadata
}
\description{
Clean PolyA Metadata by Removing Duplicate and Positional Columns
}
