% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Visualization.R
\docType{data}
\name{th}
\alias{th}
\title{Custom ggprism Theme - Deoxygenated Style}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 68.
}
\usage{
th
}
\description{
A customized ggplot2 theme based on ggprism with grey facet strips,
bold text, and enhanced borders for publication-quality figures.
}
\details{
Theme specifications:
- Base size: 11pt
- Base line size: 0.25
- Strip background: grey95 with black border
- Strip text: bold, black, size 11
- Panel spacing: 0.5 lines
- Panel border: black, 0.5 linewidth
- Legend title: size 11
- Legend text: size 12
}
\examples{
\dontrun{
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  facet_wrap(~cyl) +
  th
}


}
\keyword{datasets}
